+++ spec/rubocop/cop/style/numeric_literals_spec.rb
@@ -35,6 +35,13 @@ module Rubocop
                                'b = 123.456'])
           expect(cop.messages).to be_empty
         end
+
+        it 'ignores non-decimal literals' do
+          inspect_source(cop, ['a = 0b1010101010101',
+                               'b = 01919191919191',
+                               'c = 0xab11111111bb'])
+          expect(cop.offences).to be_empty
+        end
       end
     end
   end