+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bugs fixed
 
 * [#496](https://github.com/bbatsov/rubocop/issues/496) - Fix corner case crash in `AlignHash` cop: single key/value pair when configuration is `table` for '=>' and `separator` for `:`.
+* [#502](https://github.com/bbatsov/rubocop/issues/502) - Don't check non-decimal literals with `NumericLiterals`
 
 ## 0.13.1 (19/09/2013)